# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Data Analyst Agent API
  version: 1.0.0
  description: |
    AI Data Analyst Agent 前后端共享契约。
    具体通用规则见 API_INTEGRATION_STANDARD.md。
  contact:
    name: AI Data Analyst Agent 项目组
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发
  - url: /api/v1
    description: 同源部署
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: 查询当前用户可见项目
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
      responses:
        "200":
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Projects]
      operationId: createProject
      summary: 创建项目
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
            example:
              name: 客户流失分析
              description: 分析订阅客户的 30 日流失风险
              timezone: Asia/Shanghai
              language: zh-CN
      responses:
        "201":
          description: 项目创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: getProject
      summary: 获取项目详情
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: 更新项目元数据
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: 更新后的项目
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [Projects]
      operationId: archiveProject
      summary: 归档项目
      responses:
        "204":
          description: 已归档
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/runs/{run_id}/artifacts:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Evidence]
      operationId: listArtifacts
      summary: 查询运行产物
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: type
          in: query
          schema:
            type: string
            enum: [metric, table, chart, model, file, log, comparison]
      responses:
        "200":
          description: 产物列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactPage"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/artifacts/{artifact_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ArtifactId"
    get:
      tags: [Evidence]
      operationId: getArtifact
      summary: 获取产物元数据与展示内容
      responses:
        "200":
          description: 产物详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/artifacts/{artifact_id}/download:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ArtifactId"
    post:
      tags: [Evidence]
      operationId: createArtifactDownload
      summary: 创建有时效下载地址
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 下载信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Download"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/artifacts/{artifact_id}/file:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ArtifactId"
    get:
      tags: [Evidence]
      operationId: downloadArtifactFile
      summary: 使用有时效签名下载 Artifact 文件
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifact 文件内容
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/runs/{run_id}/claims:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Evidence]
      operationId: listClaims
      summary: 查询运行结论
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: validation_status
          in: query
          schema:
            type: string
            enum: [pending, passed, failed]
      responses:
        "200":
          description: 结论列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimPage"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/claims/{claim_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ClaimId"
    get:
      tags: [Evidence]
      operationId: getClaim
      summary: 获取结论与证据链
      responses:
        "200":
          description: 结论详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
...
  /jobs/{job_id}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [Jobs]
      operationId: getJob
      summary: 获取异步任务状态
      responses:
        "200":
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /jobs/{job_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: 取消异步任务
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 取消后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: datatrace_session
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^prj_[A-Za-z0-9]+$"
...
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^run_[A-Za-z0-9]+$"
    ArtifactId:
      name: artifact_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^art_[A-Za-z0-9]+$"
    ClaimId:
      name: claim_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^claim_[A-Za-z0-9]+$"
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^job_[A-Za-z0-9]+$"
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: field:asc 或 field:desc
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源 revision，格式为双引号包围的整数
      schema:
        type: string
        example: '"3"'
  responses:
    Unauthorized:
      description: 登录凭证缺失、无效或已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在或无权访问
              request_id: req_01JABC
              retryable: false
              details: {}
    Conflict:
      description: 当前状态或重复请求冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreconditionFailed:
      description: 客户端资源版本已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VERSION_CONFLICT
              message: 资源已被修改，请刷新后重试
              request_id: req_01JABC
              retryable: false
              details:
                current_revision: 4
    ValidationError:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ProjectCreate:
      type: object
      additionalProperties: false
      required: [name, timezone, language]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 2000
        timezone:
          type: string
          example: Asia/Shanghai
        language:
          type: string
          enum: [zh-CN, en-US]
    ProjectUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 2000
        timezone:
          type: string
        language:
          type: string
          enum: [zh-CN, en-US]
    Project:
      type: object
      required:
        - project_id
        - name
        - timezone
        - language
        - status
        - revision
        - created_at
        - updated_at
      properties:
        project_id:
          type: string
          example: prj_01JABC
        name:
          type: string
        description:
          type: [string, "null"]
        timezone:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [active, archived]
        current_dataset_version_id:
          type: [string, "null"]
        revision:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Project"
...
    Artifact:
      type: object
      required:
        - artifact_id
        - project_id
        - run_id
        - dataset_version_id
        - type
        - producer
        - status
        - checksum
        - created_at
      properties:
        artifact_id:
          type: string
          example: art_01JABC
        project_id:
          type: string
        run_id:
          type: [string, "null"]
        dataset_version_id:
          type: string
        type:
          type: string
          enum: [metric, table, chart, model, file, log, comparison]
        name:
          type: string
        producer:
          type: string
        producer_version:
          type: string
        status:
          type: string
          enum: [creating, ready, failed, expired]
        parameters:
          type: object
          additionalProperties: true
        result:
          type: [object, array, string, number, boolean, "null"]
        preview:
          type: [object, array, string, number, boolean, "null"]
        checksum:
          type: string
        downloadable:
          type: boolean
        created_at:
          type: string
          format: date-time
    ArtifactPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Artifact"
    Claim:
      type: object
      required:
        - claim_id
        - project_id
        - run_id
        - dataset_version_id
        - text
        - level
        - evidence_ids
        - limitations
        - validation_status
        - publication_status
        - created_at
      properties:
        claim_id:
          type: string
          example: claim_01JABC
        project_id:
          type: string
        run_id:
          type: string
        dataset_version_id:
          type: string
        text:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 5
          description: 1事实、2统计推断、3模型解释、4业务推测、5行动建议
        evidence_ids:
          type: array
          minItems: 1
          items:
            type: string
        limitations:
          type: array
          items:
            type: string
        validation_status:
          type: string
          enum: [pending, passed, failed]
        validation_messages:
          type: array
          items:
            type: string
        publication_status:
          type: string
          enum: [draft, published, withdrawn]
        created_at:
          type: string
          format: date-time
    ClaimPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Claim"
...
    Download:
      type: object
      required: [download_url, file_name, expires_at]
      properties:
        download_url:
          type: string
          format: uri
        file_name:
          type: string
        content_type:
          type: string
        size_bytes:
          type: [integer, "null"]
          minimum: 0
        expires_at:
          type: string
          format: date-time
    Job:
      type: object
      required:
        - job_id
        - kind
        - status
        - progress
        - created_at
        - updated_at
      properties:
        job_id:
          type: string
          example: job_01JABC
        kind:
          type: string
          enum:
            - dataset_ingestion
            - version_comparison
            - quality_scan
            - cleaning_preview
            - cleaning_execute
            - analysis_run
            - report_export
        status:
          type: string
          enum: [queued, running, cancelling, cancelled, blocked, succeeded, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        current_step:
          type: [string, "null"]
        resource_type:
          type: [string, "null"]
        resource_id:
          type: [string, "null"]
        retry_after_ms:
          type: [integer, "null"]
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error:
          anyOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: "null"
    PageMeta:
      type: object
      required: [page, page_size, total, has_more]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable, details]
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true